basic styling, moved components out of separate folders, installed packages
