+++ client/src/main.jsx
@@ -1,5 +1,6 @@
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import './styles.css'
 
 import App from './App.jsx';
 import Home from './pages/Home';
