+++ client/src/pages/Home.jsx
@@ -1,21 +1,14 @@
 import { useQuery } from '@apollo/client';
-
-// import ProfileList from '../components/ProfileList';
-
-import { QUERY_PROFILES } from '../utils/queries';
+import { QUERY_PROFILES } from '../utils/queries'
 
 const Home = () => {
 
-
   return (
     <main>
-      <div>
-        <div>
-          <h2> Hello World! </h2>
-        </div>
-      </div>
+      <h2> Hello World! </h2>
+      
     </main>
-  );
-};
+  )
+}
 
 export default Home;
