+++ client/src/styles.css
@@ -1,0 +1,33 @@
+:root {
+  --dark: #252524;
+  --light: #e7ecef;
+  --accent: #ac2290;
+  --accent-dark: #6e1b6e;
+}
+
+body {
+  background-color: var(--dark);
+  color: var(--light);
+}
+
+button,
+.btn-link {
+  background-color: var(--accent);
+  color: var(--dark);
+  padding: .25rem .75rem;
+  text-decoration: none;
+  font-size: 1rem;
+  font-weight: 700;
+  border-radius: 5px;
+  transition: all 0.3s ease-in-out;
+  border: 2px solid var(--accent);
+  &:hover {
+      background-color: var(--dark);
+      color: var(--accent);
+      border: 2px solid var(--accent);
+  }
+  &:active {
+  background-color: var(--accent-dark);
+  color: var(--accent);
+  }
+}
