+++ server/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@apollo/server": "^4.7.1",
     "bcrypt": "^5.0.0",
+    "cloudinary": "^1.41.0",
     "express": "^4.17.2",
     "graphql": "^16.6.0",
     "jsonwebtoken": "^8.5.1",
